feat: refactor database schema by removing schema.ts, updating drizzle.config, and adding quotas and apiKeys tables
