+++ src/db/schema/index.ts
@@ -1,0 +1,3 @@
+export * from "./projects";
+export * from "./events";
+export * from "./users";
